third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/13.1-expected.txt

PASS !!Function("'use strict'; function f(x,y) {}") is true
PASS !!Function("function f(x,y) {}") is true
PASS true === true
PASS Function("'use strict'; function f(x,x) {}") threw exception of type SyntaxError.
PASS !!Function("function f(x,x) {}") is true
PASS true === true
PASS Function("'use strict'; function f(x,y,z,y) {}") threw exception of type SyntaxError.
PASS !!Function("function f(x,y,z,y) {}") is true
PASS true === true
PASS Function("'use strict'; function f(a,b,c,d,e,f,g,h,d) {}") threw exception of type SyntaxError.
PASS !!Function("function f(a,b,c,d,e,f,g,h,d) {}") is true
PASS true === true
PASS !!Function("'use strict'; function f([x,y]) {}") is true
PASS !!Function("function f([x,y]) {}") is true
PASS true === true
PASS Function("'use strict'; function f([x,x]){}") threw exception of type SyntaxError.
PASS Function("function f([x,x]){}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f(x,[x]){}") threw exception of type SyntaxError.
PASS Function("function f(x,[x]){}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f(x,x) { \"use strict\" };") threw exception of type SyntaxError.
PASS Function("function f(x,x) { \"use strict\" };") threw exception of type SyntaxError.
PASS true === true
PASS true === true
PASS true === true
PASS 'use strict'; Function("x","x","'use strict'") threw exception of type SyntaxError.
PASS Function("x","x","'use strict'") threw exception of type SyntaxError.
PASS true === true
PASS true === true
PASS Function("'use strict'; (function (x,x) 2)") threw exception of type SyntaxError.
PASS Function("(function (x,x) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function (x,y) 2)") threw exception of type SyntaxError.
PASS Function("(function (x,y) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f(eval){}") threw exception of type SyntaxError.
PASS !!Function("function f(eval){}") is true
PASS true === true
PASS Function("'use strict'; function f([eval]){}") threw exception of type SyntaxError.
PASS !!Function("function f([eval]){}") is true
PASS true === true
PASS Function("'use strict'; function f({x:eval}){}") threw exception of type SyntaxError.
PASS !!Function("function f({x:eval}){}") is true
PASS true === true
PASS Function("'use strict'; function eval(){}") threw exception of type SyntaxError.
PASS !!Function("function eval(){}") is true
PASS true === true
PASS Function("'use strict'; function f(eval){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function f(eval){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f([eval]){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function f([eval]){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f({x:eval}){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function f({x:eval}){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function eval(){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function eval(){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f(eval){})") threw exception of type SyntaxError.
PASS !!Function("(function f(eval){})") is true
PASS true === true
PASS Function("'use strict'; (function f([eval]){})") threw exception of type SyntaxError.
PASS !!Function("(function f([eval]){})") is true
PASS true === true
PASS Function("'use strict'; (function f({x:eval}){})") threw exception of type SyntaxError.
PASS !!Function("(function f({x:eval}){})") is true
PASS true === true
PASS Function("'use strict'; (function eval(){})") threw exception of type SyntaxError.
PASS !!Function("(function eval(){})") is true
PASS true === true
PASS Function("'use strict'; (function f(eval){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function f(eval){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f([eval]){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function f([eval]){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f({x:eval}){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function f({x:eval}){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function eval(){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function eval(){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f(eval) 2)") threw exception of type SyntaxError.
PASS Function("(function f(eval) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f([eval]) 2)") threw exception of type SyntaxError.
PASS Function("(function f([eval]) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f({x:eval}) 2)") threw exception of type SyntaxError.
PASS Function("(function f({x:eval}) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function eval() 2)") threw exception of type SyntaxError.
PASS Function("(function eval() 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; ({set x(eval){}})") threw exception of type SyntaxError.
PASS !!Function("({set x(eval){}})") is true
PASS true === true
PASS Function("'use strict'; ({set x([eval]){}})") threw exception of type SyntaxError.
PASS !!Function("({set x([eval]){}})") is true
PASS true === true
PASS Function("'use strict'; ({set x({x:eval}){}})") threw exception of type SyntaxError.
PASS !!Function("({set x({x:eval}){}})") is true
PASS true === true
PASS Function("'use strict'; ({set x(eval){\"use strict\";}})") threw exception of type SyntaxError.
PASS Function("({set x(eval){\"use strict\";}})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; ({set x([eval]){\"use strict\";}})") threw exception of type SyntaxError.
PASS Function("({set x([eval]){\"use strict\";}})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; ({set x({x:eval}){\"use strict\";}})") threw exception of type SyntaxError.
PASS Function("({set x({x:eval}){\"use strict\";}})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f(arguments){}") threw exception of type SyntaxError.
PASS !!Function("function f(arguments){}") is true
PASS true === true
PASS Function("'use strict'; function f([arguments]){}") threw exception of type SyntaxError.
PASS !!Function("function f([arguments]){}") is true
PASS true === true
PASS Function("'use strict'; function f({x:arguments}){}") threw exception of type SyntaxError.
PASS !!Function("function f({x:arguments}){}") is true
PASS true === true
PASS Function("'use strict'; function arguments(){}") threw exception of type SyntaxError.
PASS !!Function("function arguments(){}") is true
PASS true === true
PASS Function("'use strict'; function f(arguments){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function f(arguments){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f([arguments]){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function f([arguments]){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function f({x:arguments}){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function f({x:arguments}){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; function arguments(){\"use strict\";}") threw exception of type SyntaxError.
PASS Function("function arguments(){\"use strict\";}") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f(arguments){})") threw exception of type SyntaxError.
PASS !!Function("(function f(arguments){})") is true
PASS true === true
PASS Function("'use strict'; (function f([arguments]){})") threw exception of type SyntaxError.
PASS !!Function("(function f([arguments]){})") is true
PASS true === true
PASS Function("'use strict'; (function f({x:arguments}){})") threw exception of type SyntaxError.
PASS !!Function("(function f({x:arguments}){})") is true
PASS true === true
PASS Function("'use strict'; (function arguments(){})") threw exception of type SyntaxError.
PASS !!Function("(function arguments(){})") is true
PASS true === true
PASS Function("'use strict'; (function f(arguments){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function f(arguments){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f([arguments]){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function f([arguments]){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f({x:arguments}){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function f({x:arguments}){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function arguments(){\"use strict\";})") threw exception of type SyntaxError.
PASS Function("(function arguments(){\"use strict\";})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f(arguments) 2)") threw exception of type SyntaxError.
PASS Function("(function f(arguments) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f([arguments]) 2)") threw exception of type SyntaxError.
PASS Function("(function f([arguments]) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function f({x:arguments}) 2)") threw exception of type SyntaxError.
PASS Function("(function f({x:arguments}) 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; (function arguments() 2)") threw exception of type SyntaxError.
PASS Function("(function arguments() 2)") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; ({set x(arguments){}})") threw exception of type SyntaxError.
PASS !!Function("({set x(arguments){}})") is true
PASS true === true
PASS Function("'use strict'; ({set x([arguments]){}})") threw exception of type SyntaxError.
PASS !!Function("({set x([arguments]){}})") is true
PASS true === true
PASS Function("'use strict'; ({set x({x:arguments}){}})") threw exception of type SyntaxError.
PASS !!Function("({set x({x:arguments}){}})") is true
PASS true === true
PASS Function("'use strict'; ({set x(arguments){\"use strict\";}})") threw exception of type SyntaxError.
PASS Function("({set x(arguments){\"use strict\";}})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; ({set x([arguments]){\"use strict\";}})") threw exception of type SyntaxError.
PASS Function("({set x([arguments]){\"use strict\";}})") threw exception of type SyntaxError.
PASS true === true
PASS Function("'use strict'; ({set x({x:arguments}){\"use strict\";}})") threw exception of type SyntaxError.
PASS Function("({set x({x:arguments}){\"use strict\";}})") threw exception of type SyntaxError.
PASS true === true
PASS true === true
PASS 'use strict'; Function("eval","'use strict';") threw exception of type SyntaxError.
PASS Function("eval","'use strict';") threw exception of type SyntaxError.
PASS true === true
PASS true === true
PASS 'use strict'; Function("arguments","'use strict';") threw exception of type SyntaxError.
PASS Function("arguments","'use strict';") threw exception of type SyntaxError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
